+++ constable/hash.h
@@ -11,15 +11,15 @@
 
 struct hash_ent_s {
 	struct hash_ent_s *next;
-	u_int32_t	key;
+	uintptr_t	key;
 };
 
 struct hash_s {
 	struct hash_ent_s *table[256];
 };
 
-void hash_add( struct hash_s *h, struct hash_ent_s *e, u_int32_t key );
-struct hash_ent_s *hash_find( struct hash_s *h, u_int32_t key );
+void hash_add( struct hash_s *h, struct hash_ent_s *e, uintptr_t key );
+struct hash_ent_s *hash_find( struct hash_s *h, uintptr_t key );
 
 
 #endif /* _HASH_H */
